requirements.txt
#django
django==1.4.5

#bootstrap
django-bootstrap-toolkit==2.15.0

#rest_framework(later may use it)
djangorestframework==2.3.6

#django-userena (2.0 may use it as user app)
#django-guardian (user permmition)  

#South

#django-avatar-1.0.5


#easy-thumbnails
easy-thumbnails==1.3


#PIL
PIL==1.1.7
#Pillow 


#django-simple-captcha
django-simple-captcha==0.4.0

#
Markdown==2.1.1


django-pagination

#taggit
django-taggit

django-taggit-templatetags
